refactor: un-abstract all-tab-contents
