--- src/components/command-bar/commands/search/unified-search/components/all-tab-contents/all-tab-contents.module.css
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-.divider {
-	margin-bottom: 16px;
-	margin-top: 16px;
-}
