+++ src/components/command-bar/commands/search/unified-search/components/dialog-contents/dialog-contents.module.css
@@ -13,3 +13,8 @@
 .suggestedPagesWrapper {
 	padding: 16px;
 }
+
+.divider {
+	margin-bottom: 16px;
+	margin-top: 16px;
+}
